+++ Changelog.md
@@ -1,8 +1,15 @@
-#### v1.4.11 – Released on November X, 2025
+#### v1.4.11 – Released on November 25, 2025
 
 ##### Fixed
 
 - Enhanced sidepanel language controls with responsive layout
+- Fixed various issues in Firefox
+- Fixed minor issues
+
+##### Changed
+
+- Changed default macOS shortcut to Command+A for Select Element activation
+- Disabled `Thinking Mode` by default for Gemini models to improve translation speed
 
 ---